+++ lib/dash/dash_parser.js
@@ -10,7 +10,6 @@ goog.require('goog.asserts');
 goog.require('goog.Uri');
 goog.require('shaka.abr.Ewma');
 goog.require('shaka.dash.ContentProtection');
-goog.require('shaka.dash.JsonUtils');
 goog.require('shaka.dash.MpdUtils');
 goog.require('shaka.dash.SegmentBase');
 goog.require('shaka.dash.SegmentList');
@@ -442,13 +441,8 @@ shaka.dash.DashParser = class {
     const uriObj = new goog.Uri(response.uri);
     this.lastManifestQueryParams_ = uriObj.getQueryData().toString();
 
-    const contentType = response.headers['content-type'];
-    const mimeType =
-        contentType ? contentType.toLowerCase().split(';').shift() : '';
-
     // This may throw, but it will result in a failed promise.
-    await this.parseManifest_(
-        response.data, response.uri, rootElement, mimeType);
+    await this.parseManifest(response.data, response.uri, rootElement);
     // Keep track of how long the longest manifest update took.
     const endTime = Date.now();
     const updateDuration = (endTime - startTime) / 1000.0;
@@ -468,28 +462,23 @@ shaka.dash.DashParser = class {
    * @param {string} finalManifestUri The final manifest URI, which may
    *   differ from this.manifestUri_ if there has been a redirect.
    * @param {string} rootElement MPD or Patch, depending on context
-   * @param {string} mimeType
    * @return {!Promise}
-   * @private
    */
-  async parseManifest_(data, finalManifestUri, rootElement, mimeType) {
-    let mpd;
-    if (mimeType === 'application/dash+json') {
-      const jsonString = shaka.util.StringUtils.fromBytesAutoDetect(data);
-      try {
-        const json = /** @type {!Object} */ (JSON.parse(jsonString));
-        const xmlString = shaka.dash.JsonUtils.jsonToMpd(json);
-        mpd = shaka.util.TXml.parseXmlString(xmlString, rootElement);
-      } catch (e) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MANIFEST,
-            shaka.util.Error.Code.DASH_INVALID_JSON,
-            finalManifestUri);
-      }
-    } else {
-      mpd = shaka.util.TXml.parseXml(data, rootElement);
-    }
+  parseManifest(data, finalManifestUri, rootElement) {
+    const mpd = shaka.util.TXml.parseXml(data, rootElement);
+    return this.processParsedMpd(mpd, finalManifestUri, rootElement);
+  }
+
+  /**
+   * Common point for processing an already parsed MPD (whether it comes from
+   * direct XML or from JSON converted to XML by the subclass).
+   *
+   * @param {?shaka.extern.xml.Node} mpd
+   * @param {string} finalManifestUri
+   * @param {string} rootElement
+   * @return {!Promise}
+   */
+  async processParsedMpd(mpd, finalManifestUri, rootElement) {
     if (!mpd) {
       throw new shaka.util.Error(
           shaka.util.Error.Severity.CRITICAL,
@@ -504,15 +493,12 @@ shaka.dash.DashParser = class {
     if (manifestPreprocessorTXml != defaultManifestPreprocessorTXml) {
       manifestPreprocessorTXml(mpd);
     }
-
     if (rootElement === 'Patch') {
       return this.processPatchManifest_(mpd);
     }
-
     if (!shaka.dash.MpdUtils.hasXlinks(mpd)) {
       return this.processManifest_(mpd, finalManifestUri);
     }
-
     // Process the mpd to account for xlink connections.
     const failGracefully = this.config_.dash.xlinkFailGracefully;
     const xlinkOperation = shaka.dash.MpdUtils.processXlinks(
@@ -522,7 +508,6 @@ shaka.dash.DashParser = class {
     const finalMpd = await xlinkOperation.promise;
     return this.processManifest_(finalMpd, finalManifestUri);
   }
-
 
   /**
    * Takes a formatted MPD and converts it into a manifest.
@@ -3738,5 +3723,3 @@ shaka.media.ManifestParser.registerParserByMime(
     'application/dash+xml', () => new shaka.dash.DashParser());
 shaka.media.ManifestParser.registerParserByMime(
     'video/vnd.mpeg.dash.mpd', () => new shaka.dash.DashParser());
-shaka.media.ManifestParser.registerParserByMime(
-    'application/dash+json', () => new shaka.dash.DashParser());
