+++ lib/dash/json_utils.js
@@ -1,6 +1,6 @@
 /*! @license
  * Shaka Player
- * Copyright 2016 Google LLC
+ * Copyright 2026 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
