+++ test/dash/dash_parser_live_unit.js
@@ -572,7 +572,6 @@ describe('DashParser Live', () => {
         shaka.util.AbortableOperation.completed({
           uri: redirectedUri,
           data: manifestData,
-          headers: {},
         }));
 
     const manifest = await parser.start(originalUri, playerInterface);
